[fix][dingo-mysql] Fix bug: where in subitems exceed 20 trigger NullPointerException
