+++ dingo-common/src/main/java/io/dingodb/common/util/ByteUtils.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2021 DataCanvas
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.dingodb.common.util;
+
+import javax.xml.bind.DatatypeConverter;
+
+public class ByteUtils {
+    public static byte[] hexStringToByteArray(String hexStr) {
+        hexStr = hexStr.toUpperCase();
+        int len = hexStr.length() / 2;
+        char[] hexChars = hexStr.toCharArray();
+        byte[] bytes = new byte[len];
+        for (int i = 0; i < len; i++) {
+            int pos = i * 2;
+            bytes[i] = (byte) (charToByte(hexChars[pos]) << 4 | charToByte(hexChars[pos + 1]));
+        }
+        return bytes;
+    }
+
+    public static byte charToByte(char operand) {
+        return (byte) "0123456789ABCDEF".indexOf(operand);
+    }
+
+    public static String byteArrayToHexString(byte[] bytes) {
+        return DatatypeConverter.printHexBinary(bytes);
+    }
+}
