+++ dingo-executor/src/main/java/io/dingodb/server/executor/service/UserService.java
@@ -288,7 +288,7 @@ public class UserService implements io.dingodb.verify.service.UserService {
         for (int i = 0; i < userTd.getColumns().size(); i++) {
             ColumnDefinition column = userTd.getColumn(i);
             switch (column.getName()) {
-                case "USER":
+                case "NAME":
                     row[i] = user.getUser();
                     break;
                 case "HOST":
